+++ bssopenapi-20230930/alibabacloud_bssopenapi20230930/models/_describe_budgets_request.py
@@ -14,15 +14,15 @@ class DescribeBudgetsRequest(DaraModel):
         page_no: int = None,
         page_size: int = None,
     ):
-        # The budget name. Fuzzy search is supported.
+        # The budget name. Fuzzy match is supported.
         self.budget_name = budget_name
         # The budget type.
         self.budget_type = budget_type
         # The expiration status.
         self.expire_status = expire_status
-        # The primary marketplace ID. If this parameter is left empty, the ID of the marketplace to which the current user belongs is used by default.
+        # The level-1 marketplace ID. If this parameter is left empty, the marketplace ID of the current user is used by default.
         self.nbid = nbid
-        # Page number.
+        # The page number.
         self.page_no = page_no
         # The number of entries per page.
         self.page_size = page_size